requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
beautifulsoup4
playwright
langchain-community
faiss-cpu
langchain_nomic
